requirements.txt
asgiref==3.5.0
CacheControl==0.12.10
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
Deprecated==1.2.13
Django==3.2.12
djangorestframework==3.13.1
firebase-admin==5.2.0
gcloud==0.18.3
google-api-core==2.7.1
google-api-python-client==2.43.0
google-auth==2.6.3
google-auth-httplib2==0.1.0
google-cloud-core==2.2.3
google-cloud-firestore==2.4.0
google-cloud-storage==2.2.1
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
grpcio==1.44.0
grpcio-status==1.44.0
httplib2==0.20.4
idna==3.3
jwcrypto==1.0
msgpack==1.0.3
oauth2client==4.1.3
proto-plus==1.20.3
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.14.1
pyparsing==3.0.7
python-jwt==3.3.2
pytz==2022.1
requests==2.27.1
requests-toolbelt==0.9.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.9
wrapt==1.14.0
